Created SayBaseCommand and made SayAeiou extend it.
